netcomm: FawkesNetworkClient deadlock prevention
The internal locking has been "rewired". You can now safely call most
methods from event handlers. Up to now it was not possible for instance
to wait for incoming messages in a connection established handler. The
only crucial thing that remains is connect/disconnect. That might come
later. Old programs will continue to work, but simplifications are now
possible.

The FawkesNetworkClientHandler::deregistered() method is now gone. It
was not used anywhere in a useful manner, is only useful to cover
programming errors and causes locking havoc because of circular events.